+++ bgCurrent.cu
@@ -1,4 +1,4 @@
-__device__ float bgCur(float vm) {
+__device__ double bgCur(double vm) {
   int mNeuron = threadIdx.x + blockIdx.x * blockDim.x;
   double D = 1, iBg = 0;
   double gE, gI, gNoise;
